requirements.txt
pillow==10.4.0
pyserial==3.5
PyYAML==6.0.2
